Establish configuration settings for Deviance
Have decided (by trial and error) the best pattern for settings is for
them to be defined in a NW globals file. This will ensure settings are
available for the assertion, command and reporter and leave the user
having to configure settings in only one location.

Have updated documentation, any changes to configuration values will
have to be updated here too.