Create assert for visual regression
Larger refactor of some existing pieces here, the command now
automatically generates a diff if necessary and returns a structured
object including image data for the expected, actual and resultant diff
(where applicable).

Filename generator changed to be a single use piece that returns the
paths for all potential image locations.

Example test modified to use the new assertion and the conf extended to
include the new assertion.